[scatter-plot] updates for vite
